Security of Computer Systems./homework1/webapp/requirements.txt
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-pam==1.8.4
Werkzeug==2.0.2
